Create new layout for settings activities.
